+++ dagjose/jwe_recipient_assembler.go
@@ -16,7 +16,7 @@ type jweRecipientAssembler struct {
 
 var jweRecipientMixin = mixins.MapAssembler{TypeName: "JOSERecipient"}
 
-func (j *jweRecipientAssembler) BeginMap(sizeHint int) (ipld.MapAssembler, error) {
+func (j *jweRecipientAssembler) BeginMap(sizeHint int64) (ipld.MapAssembler, error) {
 	if j.state == maState_midValue && *j.key == "header" {
 		j.recipient.header = make(map[string]ipld.Node)
 		j.state = maState_initial
@@ -31,7 +31,7 @@ func (j *jweRecipientAssembler) BeginMap(sizeHint int) (ipld.MapAssembler, error
 	}
 	return j, nil
 }
-func (j *jweRecipientAssembler) BeginList(sizeHint int) (ipld.ListAssembler, error) {
+func (j *jweRecipientAssembler) BeginList(sizeHint int64) (ipld.ListAssembler, error) {
 	return jweRecipientMixin.BeginList(sizeHint)
 }
 func (j *jweRecipientAssembler) AssignNull() error {
@@ -51,7 +51,7 @@ func (j *jweRecipientAssembler) AssignNull() error {
 func (j *jweRecipientAssembler) AssignBool(b bool) error {
 	return jweRecipientMixin.AssignBool(b)
 }
-func (j *jweRecipientAssembler) AssignInt(i int) error {
+func (j *jweRecipientAssembler) AssignInt(i int64) error {
 	return jweRecipientMixin.AssignInt(i)
 }
 func (j *jweRecipientAssembler) AssignFloat(f float64) error {
